Updated borph to 2.6.37

+++ borph/include/linux/i2c-omap.h
@@ -3,7 +3,7 @@
 
 struct omap_i2c_bus_platform_data {
 	u32		clkrate;
-	void		(*set_mpu_wkup_lat)(struct device *dev, int set);
+	void		(*set_mpu_wkup_lat)(struct device *dev, long set);
 };
 
 #endif
